# docs/api/swagger.yaml
definitions:
  api_group_secrets.Secret:
    properties:
      FolderUID:
        example: /
        type: string
      Name:
        example: DEBUG
        type: string
      Type:
        example: int
        type: string
      UID:
        example: abc-def-ghi
        type: string
      Value:
        example: Test
        type: string
    type: object
  ordering.Order:
    properties:
      Order:
        example: true
        type: boolean
      OrderBy:
        example: ID
        type: string
    type: object
  pagination.Pagination:
    properties:
      Page:
        minimum: 1
        type: integer
      PerPage:
        minimum: 1
        type: integer
    required:
    - Page
    - PerPage
    type: object
  rqrs.Error:
    properties:
      Code:
        example: 511
        type: integer
      Description:
        example: Description
        type: string
      Message:
        example: Message
        type: string
    type: object
  secrets.CreateSecretsRS:
    properties:
      Data:
        items:
          $ref: '#/definitions/api_group_secrets.Secret'
        type: array
      Errors:
        items:
          $ref: '#/definitions/rqrs.Error'
        type: array
      Pages:
        example: 14
        type: integer
      PerPage:
        example: 20
        type: integer
      Total:
        example: 280
        type: integer
    type: object
  secrets.DeleteSecretsRS:
    properties:
      Errors:
        items:
          $ref: '#/definitions/rqrs.Error'
        type: array
      Pages:
        example: 14
        type: integer
      PerPage:
        example: 20
        type: integer
      Total:
        example: 280
        type: integer
    type: object
  secrets.Folder:
    properties:
      Name:
        example: 'Folder #1'
        type: string
      ParentUID:
        example: abc-def-ghi
        type: string
      UID:
        example: abc-def-ghi
        type: string
    type: object
  secrets.GetSecretsRQ:
    properties:
      FolderUID:
        example: abc-def-ghi
        type: string
      Order:
        items:
          $ref: '#/definitions/ordering.Order'
        type: array
      Pagination:
        $ref: '#/definitions/pagination.Pagination'
    type: object
  secrets.GetSecretsRS:
    properties:
      Errors:
        items:
          $ref: '#/definitions/rqrs.Error'
        type: array
      Folders:
        items:
          $ref: '#/definitions/secrets.Folder'
        type: array
      Pages:
        example: 14
        type: integer
      PerPage:
        example: 20
        type: integer
      Secrets:
        items:
          $ref: '#/definitions/api_group_secrets.Secret'
        type: array
      Total:
        example: 280
        type: integer
    type: object
  secrets.UpdateSecretsRS:
    properties:
      Data:
        items:
          $ref: '#/definitions/api_group_secrets.Secret'
        type: array
      Errors:
        items:
          $ref: '#/definitions/rqrs.Error'
        type: array
      Pages:
        example: 14
        type: integer
      PerPage:
        example: 20
        type: integer
      Total:
        example: 280
        type: integer
    type: object
host: api.hideout.local
info:
  contact:
    email: support@hideout.com
    name: API Support
    url: https://www.swagger.io/support
  description: API for working with secrets
  license:
    name: Private
  termsOfService: https://swagger.io/terms/
  title: Hideout API
  version: "1.0"
paths:
  /public/sitemap/:
    get:
      description: Получение sitemap
      operationId: public-get-sitemap
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            type: ""
        "404":
          description: Not Found
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            type: ""
      summary: Получение sitemap
      tags:
      - Общедоступные методы
  /secrets/:
    delete:
      description: Delete secrets
      operationId: delete-secrets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secrets.DeleteSecretsRS'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/secrets.DeleteSecretsRS'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/secrets.DeleteSecretsRS'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/secrets.DeleteSecretsRS'
      summary: Delete secrets
      tags:
      - Secrets
    patch:
      description: Update secrets
      operationId: update-secrets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secrets.UpdateSecretsRS'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/secrets.UpdateSecretsRS'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/secrets.UpdateSecretsRS'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/secrets.UpdateSecretsRS'
      summary: Update secrets
      tags:
      - Secrets
    put:
      description: Create secrets
      operationId: create-secrets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secrets.CreateSecretsRS'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/secrets.CreateSecretsRS'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/secrets.CreateSecretsRS'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/secrets.CreateSecretsRS'
      summary: Create secrets
      tags:
      - Secrets
  /secrets/list/:
    post:
      description: Getting secrets list
      operationId: list-secrets
      parameters:
      - description: Secrets data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/secrets.GetSecretsRQ'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secrets.GetSecretsRS'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/secrets.GetSecretsRS'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/secrets.GetSecretsRS'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/secrets.GetSecretsRS'
      summary: Getting secrets list
      tags:
      - Брокеры
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"